Create the STL to eMesh File Converter Block
Duplicates the template for simulation, simulation/, to store the
user's outputs.

Implements two buttons: the first one opens the file explorer and
allows the upload of the desired geometry, in STL format, for
simulation; the second one generates the eMesh file by calling
OpenFOAM's surfaceFeatures utility in the terminal.

Deletes the user's simulation files at the event of closing the
window.